commit inicial

+++ ejercicio_2.py
@@ -1,2 +1,2 @@
-numero = float(input("ingresa un numero\n")) # se lee el numero 
+numero = int(input("ingresa un numero\n")) # se lee el numero 
 print(pow(numero,2)) # luego se eleva al cuadrado y por ultimo se muestra a pantalla el resultado
